move constants to shared location; change chevron dynamically; remove useless notes
