+++ webview-ui/src/components/chat/auto-approve-menu/constants.ts
@@ -1,0 +1,79 @@
+import { ActionMetadata } from "./types"
+
+export const ACTION_METADATA: ActionMetadata[] = [
+	{
+		id: "enableAll",
+		label: "Enable all",
+		shortName: "All",
+		description: "Enable all actions.",
+		icon: "codicon-checklist",
+	},
+	{
+		id: "readFiles",
+		label: "Read project files",
+		shortName: "Read",
+		description: "Allows Cline to read files within your workspace.",
+		icon: "codicon-search",
+		subAction: {
+			id: "readFilesExternally",
+			label: "Read all files",
+			shortName: "Read (all)",
+			description: "Allows Cline to read any file on your computer.",
+			icon: "codicon-folder-opened",
+			parentActionId: "readFiles",
+		},
+	},
+	{
+		id: "editFiles",
+		label: "Edit project files",
+		shortName: "Edit",
+		description: "Allows Cline to modify files within your workspace.",
+		icon: "codicon-edit",
+		subAction: {
+			id: "editFilesExternally",
+			label: "Edit all files",
+			shortName: "Edit (all)",
+			description: "Allows Cline to modify any file on your computer.",
+			icon: "codicon-files",
+			parentActionId: "editFiles",
+		},
+	},
+	{
+		id: "executeSafeCommands",
+		label: "Execute safe commands",
+		shortName: "Safe Commands",
+		description:
+			"Allows Cline to execute safe terminal commands. If the model determines a command is potentially destructive, it will still require approval.",
+		icon: "codicon-terminal",
+		subAction: {
+			id: "executeAllCommands",
+			label: "Execute all commands",
+			shortName: "All Commands",
+			description: "Allows Cline to execute all terminal commands. Use at your own risk.",
+			icon: "codicon-terminal-bash",
+			parentActionId: "executeSafeCommands",
+		},
+	},
+	{
+		id: "useBrowser",
+		label: "Use the browser",
+		shortName: "Browser",
+		description: "Allows Cline to launch and interact with any website in a browser.",
+		icon: "codicon-globe",
+	},
+	{
+		id: "useMcp",
+		label: "Use MCP servers",
+		shortName: "MCP",
+		description: "Allows Cline to use configured MCP servers which may modify filesystem or interact with APIs.",
+		icon: "codicon-server",
+	},
+]
+
+export const NOTIFICATIONS_SETTING: ActionMetadata = {
+	id: "enableNotifications",
+	label: "Enable notifications",
+	shortName: "Notifications",
+	description: "Receive system notifications when Cline requires approval to proceed or when a task is completed.",
+	icon: "codicon-bell",
+}
